refactoring, aggiunta ProcessExecutors
